refactor: restructure ItemPage component to utilize ItemDetailsTable for improved layout and data handling, and update related imports for consistency
