+++ app/app/(app)/servers/[id]/(auth)/library/[libraryId]/[itemSlug]/page.tsx
@@ -1,12 +1,12 @@
-import { Suspense } from "react";
 import { notFound } from "next/navigation";
-import { ItemDetails } from "@/components/ItemDetails";
-import { SuspenseLoading } from "@/components/SuspenseLoading";
-import { getItemBySlug, getServer } from "@/lib/db";
+import { Container } from "@/components/Container";
 import { PageTitle } from "@/components/PageTitle";
-import { Card, CardContent } from "@/components/ui/card";
+import { Skeleton } from "@/components/ui/skeleton";
+import { Suspense } from "react";
+import { getItemBySlug, getServer } from "@/lib/db";
+import { ItemDetailsTable } from "./ItemDetailsTable";
 
-interface Props {
+interface PageProps {
   params: {
     id: string;
     libraryId: string;
@@ -18,40 +18,29 @@ interface Props {
   };
 }
 
-export default async function ItemPage({ params, searchParams }: Props) {
-  try {
-    const [itemData, server] = await Promise.all([
-      getItemBySlug(params.id, params.itemSlug, searchParams.page),
-      getServer(params.id)
-    ]);
-    
-    if (!itemData || !server) {
-      notFound();
-    }
+export default async function ItemPage({ params, searchParams }: PageProps) {
+  const [itemData, server] = await Promise.all([
+    getItemBySlug(params.id, params.libraryId, params.itemSlug, searchParams),
+    getServer(params.id),
+  ]);
 
-    return (
-      <div className="flex-1 space-y-4 p-8 pt-6">
-        <div className="flex items-center justify-between">
-          <div className="space-y-1">
-            <PageTitle title={itemData.item.name} />
-          </div>
-        </div>
-        <div className="grid gap-6">
-          <Card>
-            <CardContent className="p-6">
-              <Suspense fallback={<SuspenseLoading />}>
-                <ItemDetails 
-                  item={itemData.item} 
-                  statistics={itemData.statistics} 
-                  serverUrl={server.url}
-                />
-              </Suspense>
-            </CardContent>
-          </Card>
-        </div>
-      </div>
-    );
-  } catch (error) {
+  if (!itemData || !server) {
     notFound();
   }
+
+  return (
+    <Container>
+      <PageTitle
+        title={itemData.item.name}
+        subtitle="View detailed statistics for this item."
+      />
+      <Suspense fallback={<Skeleton className="h-[500px]" />}>
+        <ItemDetailsTable
+          item={itemData.item}
+          statistics={itemData.statistics}
+          serverUrl={server.url}
+        />
+      </Suspense>
+    </Container>
+  );
 }
